requirements.txt
aioice==0.9.0
aiortc==1.7.0
altair==5.2.0
asgiref==3.7.2
attrs==23.2.0
av==11.0.0
blinker==1.7.0
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.0
cryptography==42.0.5
cycler==0.12.1
Django==5.0.3
dnspython==2.6.1
filelock==3.13.1
Flask==3.0.0
fonttools==4.49.0
fsspec==2024.2.0
gitdb==4.0.11
GitPython==3.1.42
google-crc32c==1.5.0
idna==3.6
ifaddr==0.2.0
importlib-metadata==7.0.1
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.3
mdurl==0.1.2
mpmath==1.3.0
mysqlclient==2.2.4
networkx==3.2.1
numpy==1.26.4
opencv-python==4.9.0.80
opencv-python-headless==4.9.0.80
packaging==23.2
pandas==2.2.1
pillow==10.2.0
protobuf==4.25.3
psutil==5.9.8
py-cpuinfo==9.0.0
py-solc==3.2.0
pyarrow==15.0.0
pycparser==2.21
pydeck==0.8.1b0
pyee==11.1.0
Pygments==2.17.2
pylibsrtp==0.10.0
pyOpenSSL==24.0.0
pyparsing==3.1.1
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
referencing==0.33.0
requests==2.31.0
rich==13.7.1
rpds-py==0.18.0
scipy==1.12.0
seaborn==0.13.2
semantic-version==2.10.0
setuptools==69.0.2
six==1.16.0
smmap==5.0.1
sqlparse==0.4.4
streamlit==1.31.1
streamlit-webrtc==0.47.1
sympy==1.12
tenacity==8.2.3
thop==0.1.1.post2209072238
toml==0.10.2
toolz==0.12.1
torch==2.2.1
torchvision==0.17.1
tornado==6.4
tqdm==4.66.2
typing_extensions==4.10.0
tzdata==2024.1
tzlocal==5.2
ultralytics==8.1.23
urllib3==2.2.1
validators==0.22.0
watchdog==4.0.0
Werkzeug==3.0.1
wheel==0.41.3
zipp==3.17.0
